flexisource/requirements.txt
amqp==5.3.1
APScheduler==3.11.0
asgiref==3.8.1
asttokens==3.0.0
billiard==4.2.1
celery==5.4.0
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
decorator==5.1.1
Django==5.1
django-rest-framework==0.1.0
djangorestframework==3.15.2
executing==2.2.0
iniconfig==2.0.0
ipdb==0.13.13
ipython==8.32.0
jedi==0.19.2
kombu==5.4.2
matplotlib-inline==0.1.7
numpy==2.2.2
packaging==24.2
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pluggy==1.5.0
prompt_toolkit==3.0.50
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.1
pytest==8.3.4
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.1
redis==5.2.1
six==1.17.0
sqlparse==0.5.3
stack-data==0.6.3
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2025.1
tzlocal==5.2
vine==5.1.0
wcwidth==0.2.13
